Refactor Notification
